Moved non-standard components into local libraries.
neotron-32.lib is the symbol library.
neotron-32.pretty is the footprint library.

Also swapped the DE9 and DB25 headers for IDC boxed versions. Involved
moving a resistor out of the way.

The Keystone 103 battery holder is a straight copy of the standard
footprint, but with the Courtyard removed to bypass the 'Courtyard is
not closed' DRC error.